+++ test/lint/sets.rb
@@ -3,9 +3,17 @@
 module Lint
   module Sets
     def test_sadd
-      assert_equal true, r.sadd("foo", "s1")
-      assert_equal true, r.sadd("foo", "s2")
-      assert_equal false, r.sadd("foo", "s1")
+      assert_equal 1, r.sadd("foo", "s1")
+      assert_equal 1, r.sadd("foo", "s2")
+      assert_equal 0, r.sadd("foo", "s1")
+
+      assert_equal ["s1", "s2"], r.smembers("foo").sort
+    end
+
+    def test_sadd?
+      assert_equal true, r.sadd?("foo", "s1")
+      assert_equal true, r.sadd?("foo", "s2")
+      assert_equal false, r.sadd?("foo", "s1")
 
       assert_equal ["s1", "s2"], r.smembers("foo").sort
     end
@@ -17,12 +25,30 @@ module Lint
       assert_equal ["s1", "s2", "s3"], r.smembers("foo").sort
     end
 
+    def test_variadic_sadd?
+      assert_equal true, r.sadd?("foo", ["s1", "s2"])
+      assert_equal true, r.sadd?("foo", ["s1", "s2", "s3"])
+      assert_equal false, r.sadd?("foo", ["s1", "s2"])
+
+      assert_equal ["s1", "s2", "s3"], r.smembers("foo").sort
+    end
+
     def test_srem
       r.sadd("foo", "s1")
       r.sadd("foo", "s2")
 
-      assert_equal true, r.srem("foo", "s1")
-      assert_equal false, r.srem("foo", "s3")
+      assert_equal 1, r.srem("foo", "s1")
+      assert_equal 0, r.srem("foo", "s3")
+
+      assert_equal ["s2"], r.smembers("foo")
+    end
+
+    def test_srem?
+      r.sadd("foo", "s1")
+      r.sadd("foo", "s2")
+
+      assert_equal true, r.srem?("foo", "s1")
+      assert_equal false, r.srem?("foo", "s3")
 
       assert_equal ["s2"], r.smembers("foo")
     end
@@ -35,6 +61,18 @@ module Lint
       assert_equal 1, r.srem("foo", ["s1", "aaa"])
       assert_equal 0, r.srem("foo", ["bbb", "ccc", "ddd"])
       assert_equal 1, r.srem("foo", ["eee", "s3"])
+
+      assert_equal ["s2"], r.smembers("foo")
+    end
+
+    def test_variadic_srem?
+      r.sadd("foo", "s1")
+      r.sadd("foo", "s2")
+      r.sadd("foo", "s3")
+
+      assert_equal true, r.srem?("foo", ["s1", "aaa"])
+      assert_equal false, r.srem?("foo", ["bbb", "ccc", "ddd"])
+      assert_equal true, r.srem?("foo", "eee", "s3")
 
       assert_equal ["s2"], r.smembers("foo")
     end